# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "道路交通事故风险预估与后果预测系统 API",
    "description": "交通事故上报、风险评估、调度任务、用户管理、地图服务、指挥中心等全部后端接口",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "summary": "分页查询用户列表",
        "description": "查询所有用户，返回分页结果",
        "tags": ["系统管理"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UserResponse" }
                    },
                    "totalElements": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "number": { "type": "integer" },
                    "size": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建用户",
        "description": "创建新用户，必填：fullName、username、role、password",
        "tags": ["系统管理"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}": {
      "put": {
        "summary": "修改用户信息",
        "description": "修改用户信息（所有字段可选，只传需要修改的字段）",
        "tags": ["系统管理"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除用户",
        "description": "删除指定用户",
        "tags": ["系统管理"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功（无返回内容）"
          }
        }
      }
    },
    "/api/v1/admin/operation-logs": {
      "get": {
        "summary": "分页查询操作日志",
        "description": "查询操作审计日志",
        "tags": ["系统管理"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/OperationLog" }
                    },
                    "totalElements": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "number": { "type": "integer" },
                    "size": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "description": "用户角色",
        "enum": ["FIELD_OFFICER", "COMMAND_CENTER", "RESCUE_WORKER", "ADMIN"]
      },
      "UserStatus": {
        "type": "string",
        "description": "用户状态",
        "enum": ["ENABLED", "DISABLED"]
      },
...
      "CreateUserRequest": {
        "type": "object",
        "description": "创建用户请求",
        "required": ["fullName", "username", "role", "password"],
        "properties": {
          "fullName": { "type": "string", "description": "真实姓名（必填）", "example": "张三" },
          "username": { "type": "string", "description": "用户名（必填，唯一）", "example": "zhangsan" },
          "phone": { "type": "string", "description": "手机号", "example": "13800138000" },
          "email": { "type": "string", "format": "email", "description": "邮箱" },
          "role": { "$ref": "#/components/schemas/UserRole", "description": "角色（必填）" },
          "password": { "type": "string", "description": "密码（必填）", "example": "123456" }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "description": "修改用户请求（所有字段可选）",
        "properties": {
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "status": { "$ref": "#/components/schemas/UserStatus" },
          "password": { "type": "string" }
        }
      },
      "UserResponse": {
        "type": "object",
        "description": "用户信息响应",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "fullName": { "type": "string" },
          "username": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "status": { "$ref": "#/components/schemas/UserStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "OperationLog": {
        "type": "object",
        "description": "操作日志",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "operatorUserId": { "type": "integer", "format": "int64", "description": "操作人用户ID" },
          "operationType": { "type": "string", "description": "操作类型", "maxLength": 80 },
          "objectType": { "type": "string", "description": "操作对象类型", "maxLength": 80 },
          "objectId": { "type": "string", "description": "操作对象ID", "maxLength": 80 },
          "ipAddress": { "type": "string", "description": "IP地址", "maxLength": 80 },
          "detail": { "type": "string", "description": "操作详情", "maxLength": 1000 }
        }
      },
...
    },
...
  },
...
}